+++ src/main/java/ca/arctechlabs/aoc/y2023/models/Coordinates.java
@@ -1,30 +1,30 @@
 package ca.arctechlabs.aoc.y2023.models;
 
 public class Coordinates{
-    private long x;
-    private long y;
+    private int x;
+    private int y;
 
     public Coordinates() {
     }
 
-    public Coordinates(long x, long y) {
+    public Coordinates(int x, int y) {
         this.x = x;
         this.y = y;
     }
 
-    public void setX(long x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public void setY(long y) {
+    public void setY(int y) {
         this.y = y;
     }
 
-    public long getX() {
+    public int getX() {
         return x;
     }
 
-    public long getY() {
+    public int getY() {
         return y;
     }
